+++ Gw2Sharp.Tests/ConnectionTests.cs
@@ -115,6 +115,7 @@ namespace Gw2Sharp.Tests
         [InlineData("membership required", HttpStatusCode.Forbidden, typeof(MembershipRequiredException))]
         [InlineData("access restricted to guild leaders", HttpStatusCode.Forbidden, typeof(RestrictedToGuildLeadersException))]
         [InlineData("not found", HttpStatusCode.NotFound, typeof(NotFoundException))]
+        [InlineData("too many requests", HttpStatusCode.TooManyRequests, typeof(TooManyRequestsException))]
         [InlineData("server error", HttpStatusCode.InternalServerError, typeof(ServerErrorException))]
         [InlineData("service unavailable", HttpStatusCode.ServiceUnavailable, typeof(ServiceUnavailableException))]
         public async Task ExceptionRequestTest(string errorText, HttpStatusCode statusCode, Type exceptionType)